tests/wpt/web-platform-tests/referrer-policy/strict-origin/http-rp/same-origin/http-http/script-tag/no-redirect/insecure-protocol.http.html

<!DOCTYPE html>
<!-- DO NOT EDIT! Generated by referrer-policy/generic/tools/generate.py using referrer-policy/generic/template/test.release.html.template. -->
<html>
  <head>
    <title>Referrer-Policy: Referrer Policy is set to 'strict-origin'</title>
    <meta name="description" content="Check that non a priori insecure subresource gets only the origin portion of the referrer URL. A priori insecure subresource gets no referrer information.">
    <!-- No meta: Referrer policy delivered via HTTP headers. -->
    <link rel="author" title="Kristijan Burnik" href="burnik@chromium.org">
    <link rel="help" href="https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin">
    <meta name="assert" content="The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the http-rp
                                 delivery method with no-redirect and when
                                 the target request is same-origin.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- TODO(kristijanburnik): Minify and merge both: -->
    <script src="/referrer-policy/generic/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js?pipe=sub"></script>
  </head>
  <body>
    <script>
      ReferrerPolicyTestCase(
        {
          "referrer_policy": "strict-origin",
          "delivery_method": "http-rp",
          "redirection": "no-redirect",
          "origin": "same-origin",
          "source_protocol": "http",
          "target_protocol": "http",
          "subresource": "script-tag",
          "subresource_path": "/referrer-policy/generic/subresource/script.py",
          "referrer_url": "origin"
        },
        document.querySelector("meta[name=assert]").content,
        new SanityChecker()
      ).start();
      </script>
    <div id="log"></div>
  </body>
</html>
