Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/strict-origin/meta-referrer/same-origin/http-http/script-tag/no-redirect/insecure-protocol.http-expected.txt


PASS The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the meta-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin.
